requirements.txt
gradio
langchain-openai
langchain-core
python-dotenv
langchain